# Backend/openapi.yaml
openapi: 3.0.3
info:
  title: WealthTrack Microservice Ecosystem API
  version: 1.0.0
  description: |
    Code-derived API contract for every custom controller outside `MAIN`.

    **Authentication:** no Spring Security configuration, authentication filter, or
    authorization policy exists in the reviewed services. Every operation therefore
    declares `security: []`. The reusable 401 and 403 responses are future-security
    templates only and are not currently emitted by application authentication.

    **Headers:** JSON operations accept `Content-Type: application/json` and normally
    return `application/json`; clients may send `Accept: application/json`. Delete and
    internal command operations return no body. There are no application-specific
    request headers in the source code.

    **Gateway use:** each business path has a direct-service server and an API Gateway
    server. The gateway removes `/user`, `/bank`, `/portfolio`, `/trading`, or
    `/product` before forwarding. Internal endpoints are not access-controlled and are
    also reachable through the gateway as currently configured.

    **Uncertain/framework contracts:** Eureka exposes its library-managed dashboard and
    discovery protocol under `/eureka/**`, but the repository defines no custom Eureka
    controller or DTO contract. Those framework endpoints are described in
    `MICROSERVICE_API_MAP.md` and deliberately not invented here.
servers:
  - url: http://localhost:8081
...
security: []
paths:
...
  /api/bank-accounts:
    servers:
      - url: http://localhost:8083
        description: Bank Service direct
      - url: http://localhost:8081/bank
        description: Bank Service through API Gateway
    post:
      tags: [Bank - Accounts]
      operationId: createBankAccount
      summary: Register a bank account
      description: |
        Validates the referenced user through User Service, requires that user to be
        ACTIVE, normalizes the account number, and creates an ACTIVE account. The first
        account for a user becomes primary; requesting primary clears the user's other
        primary flags. The account number must be globally unique.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBankAccountRequest' }
            example:
              userId: 10
              bankName: HDFC Bank
              branchName: Andheri East
              accountNumber: '50100234567890'
              accountType: SAVINGS
              ifscCode: HDFC0001234
              openingBalance: 250000.00
              primaryAccount: true
      responses:
        '201':
          description: Bank account registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BankAccountResponse' }
              example:
                bankAccountId: 1
                userId: 10
                bankName: HDFC Bank
                branchName: Andheri East
                maskedAccountNumber: "**********7890"
                accountType: SAVINGS
                ifscCode: HDFC0001234
                balance: 250000.00
                primaryAccount: true
                status: ACTIVE
                createdAt: '2026-07-30T10:20:00'
                updatedAt: '2026-07-30T10:20:00'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    get:
      tags: [Bank - Accounts]
      operationId: listBankAccounts
      summary: List and filter bank accounts
      description: Filters in memory by optional user ID, parsed status, and primary flag. An invalid status string returns 400.
      security: []
      parameters:
        - name: userId
          in: query
          required: false
          schema: { type: integer, format: int64 }
          example: 10
        - name: status
          in: query
          required: false
          schema: { $ref: '#/components/schemas/BankAccountStatus' }
          example: ACTIVE
        - name: primary
          in: query
          required: false
          schema: { type: boolean }
          example: true
      responses:
        '200':
          description: Matching bank accounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BankAccountResponse' }
              example:
                - bankAccountId: 1
                  userId: 10
                  bankName: HDFC Bank
                  branchName: Andheri East
                  maskedAccountNumber: "**********7890"
                  accountType: SAVINGS
                  ifscCode: HDFC0001234
                  balance: 250000.00
                  primaryAccount: true
                  status: ACTIVE
                  createdAt: '2026-07-30T10:20:00'
                  updatedAt: '2026-07-30T10:20:00'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/bank-accounts/{id}:
    servers:
      - url: http://localhost:8083
        description: Bank Service direct
      - url: http://localhost:8081/bank
        description: Bank Service through API Gateway
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Bank - Accounts]
      operationId: getBankAccountById
      summary: Get a bank account
      description: Returns account metadata, masked account number, and current balance.
      security: []
      responses:
        '200':
          description: Bank account found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BankAccountResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    put:
      tags: [Bank - Accounts]
      operationId: updateBankAccount
      summary: Update mutable bank account fields
      description: Updates only non-null bank name, branch, account type, IFSC, and status fields. Setting CLOSED also clears the primary flag.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateBankAccountRequest' }
            example:
              bankName: HDFC Bank
              branchName: Bandra Kurla Complex
              accountType: SAVINGS
              ifscCode: HDFC0004321
              status: ACTIVE
      responses:
        '200':
          description: Updated bank account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BankAccountResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    delete:
      tags: [Bank - Accounts]
      operationId: closeBankAccount
      summary: Close a bank account
      description: Soft-closes the account by setting status CLOSED and primaryAccount false; it does not delete the database row.
      security: []
      responses:
        '204': { description: Account closed; no response body }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/bank-accounts/{id}/primary:
    servers:
      - url: http://localhost:8083
        description: Bank Service direct
      - url: http://localhost:8081/bank
        description: Bank Service through API Gateway
    patch:
      tags: [Bank - Accounts]
      operationId: makeBankAccountPrimary
      summary: Make an account primary
      description: Marks an ACTIVE account primary and clears the primary flag from the user's other accounts. No request body is accepted.
      security: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Updated primary account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BankAccountResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/bank-accounts/{id}/debit:
    servers:
      - url: http://localhost:8083
        description: Bank Service direct
      - url: http://localhost:8081/bank
        description: Bank Service through API Gateway
    post:
      tags: [Bank - Accounts]
      operationId: debitBankAccount
      summary: Debit a bank account
      description: |
        Locks the account row and subtracts the amount when the account is ACTIVE and
        funded. Business rejection (inactive account or insufficient funds) still
        returns HTTP 200 with `approved: false`. Trading Service consumes this endpoint.
        The transaction reference is returned but is not persisted or deduplicated.
      security: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DebitRequest' }
            example: { amount: 1450.00, transactionReference: TRADE-8 }
      responses:
        '200':
          description: Debit decision
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DebitResult' }
              examples:
                approved:
                  value: { approved: true, reference: TRADE-8, failureReason: null }
                rejected:
                  value: { approved: false, reference: TRADE-8, failureReason: Insufficient balance }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/bank-accounts/{id}/credit:
    servers:
      - url: http://localhost:8083
        description: Bank Service direct
      - url: http://localhost:8081/bank
        description: Bank Service through API Gateway
    post:
      tags: [Bank - Accounts]
      operationId: creditBankAccount
      summary: Credit a bank account
      description: |
        Locks the account row and adds the amount unless the account is CLOSED.
        A CLOSED-account rejection is HTTP 200 with `successful: false`. Trading uses
        this for SELL proceeds and BUY compensation. References are not deduplicated.
      security: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreditRequest' }
            example: { amount: 1450.00, transactionReference: REVERSAL-TRADE-8 }
      responses:
        '200':
          description: Credit result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreditResult' }
              examples:
                successful:
                  value: { successful: true, reference: REVERSAL-TRADE-8, failureReason: null }
                rejected:
                  value: { successful: false, reference: TRADE-8, failureReason: Cannot credit a CLOSED bank account }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
  /api/trade-transactions:
    servers:
      - url: http://localhost:8085
        description: Trading Service direct
      - url: http://localhost:8081/trading
        description: Trading Service through API Gateway
    post:
      tags: [Trading - Transactions]
      operationId: createTradeTransaction
      summary: Execute a BUY or SELL trade
      description: |
        Persists PENDING, obtains Product Service's active currentPrice (the request's
        unitPrice is therefore overwritten), validates the Portfolio, then runs a
        synchronous saga:

        * BUY: Bank debit -> Portfolio apply; credit compensation on Portfolio failure.
        * SELL: Portfolio apply -> Bank credit; reverse Portfolio command on credit failure.

        A completed trade returns 201. Any caught execution failure is persisted as
        FAILED and returned as 422 with the same TradeTransactionResponse shape and a
        failureReason. The external calls and compensations are not globally atomic.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTradeTransactionRequest' }
            example:
              portfolioAccountId: 1
              holdingId: 1
              productId: 1
              bankAccountId: 1
              transactionType: BUY
              quantity: 1.00
              unitPrice: 1450.00
      responses:
        '201':
          description: Trade completed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TradeTransactionResponse' }
              example:
                transactionId: 8
                portfolioAccountId: 1
                holdingId: 1
                productId: 1
                transactionType: BUY
                quantity: 1.00
                unitPrice: 1450.00
                totalAmount: 1450.00
                transactionStatus: COMPLETED
                transactionDate: '2026-07-30T11:00:00'
                failureReason: null
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422':
          description: Trade was accepted for execution but failed; this is a persisted business result, not the standard error envelope
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TradeTransactionResponse' }
              example:
                transactionId: 9
                portfolioAccountId: 1
                holdingId: 1
                productId: 1
                transactionType: BUY
                quantity: 1.00
                unitPrice: 1450.00
                totalAmount: 1450.00
                transactionStatus: FAILED
                transactionDate: '2026-07-30T11:05:00'
                failureReason: Insufficient balance
        '500': { $ref: '#/components/responses/InternalServerError' }
    get:
      tags: [Trading - Transactions]
      operationId: listTradeTransactions
      summary: List and filter trades
      description: Filters stored transactions by optional portfolio ID, case-insensitive status/type strings, and inclusive transactionDate boundaries. Unknown status/type values simply produce no matching rows.
      security: []
      parameters:
        - name: portfolioAccountId
          in: query
          schema: { type: integer, format: int64 }
          example: 1
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/TransactionStatus' }
          example: COMPLETED
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/TransactionType' }
          example: BUY
        - name: startDate
          in: query
          description: Inclusive ISO-8601 local date-time; no timezone offset is modeled by LocalDateTime.
          schema: { type: string, format: date-time }
          example: '2026-07-01T00:00:00'
        - name: endDate
          in: query
          description: Inclusive ISO-8601 local date-time.
          schema: { type: string, format: date-time }
          example: '2026-07-31T23:59:59'
      responses:
        '200':
          description: Matching trades
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TradeTransactionResponse' }
              example:
                - transactionId: 8
                  portfolioAccountId: 1
                  holdingId: 1
                  productId: 1
                  transactionType: BUY
                  quantity: 1.00
                  unitPrice: 1450.00
                  totalAmount: 1450.00
                  transactionStatus: COMPLETED
                  transactionDate: '2026-07-30T11:00:00'
                  failureReason: null
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/trade-transactions/{id}:
    servers:
      - url: http://localhost:8085
        description: Trading Service direct
      - url: http://localhost:8081/trading
        description: Trading Service through API Gateway
    get:
      tags: [Trading - Transactions]
      operationId: getTradeTransaction
      summary: Get a trade
      description: Returns one stored trade. failureReason is not persisted by the entity and is therefore null when a previously failed trade is read back.
      security: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Trade found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TradeTransactionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: Resource identifier
      schema: { type: integer, format: int64, minimum: 1 }
      example: 1
...
  responses:
    BadRequest:
      description: Request validation, enum parsing, date-range, or business-rule failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StandardApiError' }
          examples:
            validation:
              summary: Validation error
              value:
                timestamp: '2026-07-30T11:15:00'
                status: 400
                error: Bad Request
                message: Request validation failed
                path: /api/bank-accounts
                fieldErrors:
                  accountNumber: must not be blank
            businessRule:
              summary: Business-rule error
              value:
                timestamp: '2026-07-30T11:15:00'
                status: 400
                message: Portfolio account 1 is not active
    Unauthorized:
      description: |
        Reusable 401 template. No reviewed service currently authenticates requests, so
        this response is not emitted by application security in the current code.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StandardApiError' }
          example:
            timestamp: '2026-07-30T11:15:00'
            status: 401
            error: Unauthorized
            message: Authentication is required
            path: /api/example
    Forbidden:
      description: |
        Reusable 403 template. No reviewed service currently enforces roles or
        permissions, so this response is not emitted by application authorization.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StandardApiError' }
          example:
            timestamp: '2026-07-30T11:15:00'
            status: 403
            error: Forbidden
            message: The authenticated principal cannot perform this operation
            path: /api/example
    NotFound:
      description: Requested resource does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StandardApiError' }
          example:
            timestamp: '2026-07-30T11:15:00'
            status: 404
            error: Not Found
            message: Resource not found
            path: /api/example/999
    Conflict:
      description: Duplicate or data-integrity conflict; only handlers that map conflicts return this status
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StandardApiError' }
          example:
            timestamp: '2026-07-30T11:15:00'
            status: 409
            error: Conflict
            message: The record conflicts with existing data
            path: /api/example
...
    InternalServerError:
      description: |
        Unexpected server or dependency failure. Bank and Portfolio define this envelope
        explicitly. User, Product, Trading, and Gateway may fall through to Spring's
        default error handling, so their exact 500 body is not fully determined.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StandardApiError' }
          example:
            timestamp: '2026-07-30T11:15:00'
            status: 500
            error: Internal Server Error
            message: Internal server error
            path: /api/example
...
  schemas:
    StandardApiError:
      type: object
      description: |
        Superset reusable error envelope covering the three code-defined formats. Only
        `timestamp`, `status`, and `message` are common to all custom handlers. Optional
        fields appear in Bank/Portfolio handlers. Spring default errors can contain
        additional fields.
      required: [timestamp, status, message]
      properties:
        timestamp:
          type: string
          format: date-time
          example: '2026-07-30T11:15:00'
        status:
          type: integer
          format: int32
          example: 400
        error:
          type: string
          nullable: true
          example: Bad Request
        message:
          type: string
          example: Request validation failed
        path:
          type: string
          nullable: true
          example: /api/bank-accounts
        fieldErrors:
          type: object
          nullable: true
          additionalProperties: { type: string }
          example: { accountNumber: must not be blank }
        validationErrors:
          type: object
          nullable: true
          additionalProperties: { type: string }
          example: { quantity: must be greater than or equal to 0.0001 }
...
    AccountType:
      type: string
      enum: [SAVINGS, CURRENT, SALARY, NRE, NRO, OTHER]
    BankAccountStatus:
      type: string
      enum: [ACTIVE, INACTIVE, BLOCKED, CLOSED]
...
    CreateBankAccountRequest:
      type: object
      required: [userId, bankName, accountNumber, accountType]
      properties:
        userId: { type: integer, format: int64, minimum: 1, example: 10 }
        bankName: { type: string, minLength: 1, maxLength: 120, example: HDFC Bank }
        branchName: { type: string, maxLength: 120, nullable: true, example: Andheri East }
        accountNumber: { type: string, minLength: 1, maxLength: 50, example: '50100234567890' }
        accountType:
          description: Case-insensitive enum string parsed by the service.
          allOf: [{ $ref: '#/components/schemas/AccountType' }]
        ifscCode:
          type: string
          nullable: true
          pattern: '^[A-Za-z]{4}0[A-Za-z0-9]{6}$'
          example: HDFC0001234
        openingBalance: { type: number, minimum: 0, nullable: true, default: 0, example: 250000.00 }
        primaryAccount: { type: boolean, nullable: true, default: false, example: true }
    UpdateBankAccountRequest:
      type: object
      properties:
        bankName: { type: string, minLength: 1, maxLength: 120, nullable: true, example: HDFC Bank }
        branchName: { type: string, maxLength: 120, nullable: true, example: Bandra Kurla Complex }
        accountType:
          allOf: [{ $ref: '#/components/schemas/AccountType' }]
          nullable: true
        ifscCode:
          type: string
          nullable: true
          pattern: '^[A-Za-z]{4}0[A-Za-z0-9]{6}$'
          example: HDFC0004321
        status:
          allOf: [{ $ref: '#/components/schemas/BankAccountStatus' }]
          nullable: true
    BankAccountResponse:
      type: object
      required: [bankAccountId, userId, bankName, maskedAccountNumber, accountType, balance, primaryAccount, status, createdAt, updatedAt]
      properties:
        bankAccountId: { type: integer, format: int64, example: 1 }
        userId: { type: integer, format: int64, example: 10 }
        bankName: { type: string, example: HDFC Bank }
        branchName: { type: string, nullable: true, example: Andheri East }
        maskedAccountNumber: { type: string, example: "**********7890" }
        accountType: { $ref: '#/components/schemas/AccountType' }
        ifscCode: { type: string, nullable: true, example: HDFC0001234 }
        balance: { type: number, example: 250000.00 }
        primaryAccount: { type: boolean, example: true }
        status: { $ref: '#/components/schemas/BankAccountStatus' }
        createdAt: { type: string, format: date-time, example: '2026-07-30T10:20:00' }
        updatedAt: { type: string, format: date-time, example: '2026-07-30T10:20:00' }
    DebitRequest:
      type: object
      required: [amount, transactionReference]
      properties:
        amount: { type: number, minimum: 0.01, example: 1450.00 }
        transactionReference: { type: string, minLength: 1, example: TRADE-8 }
    CreditRequest:
      type: object
      required: [amount, transactionReference]
      properties:
        amount: { type: number, minimum: 0.01, example: 1450.00 }
        transactionReference: { type: string, minLength: 1, example: REVERSAL-TRADE-8 }
    DebitResult:
      type: object
      required: [approved, reference]
      properties:
        approved: { type: boolean, example: true }
        reference: { type: string, example: TRADE-8 }
        failureReason: { type: string, nullable: true, example: null }
    CreditResult:
      type: object
      required: [successful, reference]
      properties:
        successful: { type: boolean, example: true }
        reference: { type: string, example: TRADE-8 }
        failureReason: { type: string, nullable: true, example: null }
...
    TransactionType:
      type: string
      enum: [BUY, SELL]
    TransactionStatus:
      type: string
      enum: [PENDING, COMPLETED, FAILED, REVERSED]
    CreateTradeTransactionRequest:
      type: object
      required: [portfolioAccountId, holdingId, productId, bankAccountId, transactionType, quantity, unitPrice]
      properties:
        portfolioAccountId: { type: integer, format: int64, minimum: 1, example: 1 }
        holdingId: { type: integer, format: int64, minimum: 1, example: 1 }
        productId: { type: integer, format: int64, minimum: 1, example: 1 }
        bankAccountId: { type: integer, format: int64, minimum: 1, example: 1 }
        transactionType:
          description: Parsed case-insensitively; only BUY and SELL are accepted.
          allOf: [{ $ref: '#/components/schemas/TransactionType' }]
        quantity: { type: number, minimum: 0.01, example: 1.00 }
        unitPrice:
          type: number
          minimum: 0.01
          description: Required and validated, but execution replaces it with Product Service currentPrice.
          example: 1450.00
    TradeTransactionResponse:
      type: object
      required: [transactionId, portfolioAccountId, holdingId, productId, transactionType, quantity, unitPrice, totalAmount, transactionStatus, transactionDate]
      properties:
        transactionId: { type: integer, format: int64, example: 8 }
        portfolioAccountId: { type: integer, format: int64, example: 1 }
        holdingId: { type: integer, format: int64, example: 1 }
        productId: { type: integer, format: int64, example: 1 }
        transactionType: { $ref: '#/components/schemas/TransactionType' }
        quantity: { type: number, example: 1.00 }
        unitPrice: { type: number, example: 1450.00 }
        totalAmount: { type: number, example: 1450.00 }
        transactionStatus: { $ref: '#/components/schemas/TransactionStatus' }
        transactionDate: { type: string, format: date-time, example: '2026-07-30T11:00:00' }
        failureReason:
          type: string
          nullable: true
          description: Populated only in the immediate 422 response; the entity does not persist this field.
          example: null
...